+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - 【重要】jQuery3.6.1兼容调整，现在默认使用jQuery版本默认使用3
 - 新增 **bjui.destroyModules** 事件，用于监听dialog或navtab销毁或重载
 - 新增 **autocomplete** 插件示例
+- 新增 navtab **reloadWithSearch** 方法，添加ajax回调支持
 
 ## v1.5.1